Konfiguration über Parameter hinzugefügt

+++ 12V/12V.c
@@ -487,10 +487,10 @@ void getDataForConfigFile(int relais, void* config) {
         strpMax = readStdinLine();
         int maxRest;
         maxRest = pconfig->mcp.maxPConverter-pconfig->mcp.maxPMicroController;
-        if( (int)strpMax > (int)maxRest){
+        if((int)strpMax > (int)maxRest){
             printf("\n\e[0;31m Dieses Gerät hat einen zu hohen Watt Verbrauch. Es wird nie eingeschaltet werden können!\e[0m");
         }
-        if( strpMax < 1){
+        if((int)strpMax < (int)1){
             printf("\n\e[0;31m Dieses Gerät hat eine falsche Watt Angabe!\e[0m");
         }
